requirements.txt
beautifulsoup4>=4.9.3
requests>=2.25.1
matplotlib>=3.4.3
python-dotenv>=1.0.1
pandas>=1.3.3
